+++ crates/cli/src/utils/cmd.rs
@@ -141,29 +141,20 @@ pub fn eta_key(state: &indicatif::ProgressState, f: &mut dyn Write) {
     write!(f, "{:.1}s", state.eta().as_secs_f64()).unwrap()
 }
 
-#[macro_export]
-macro_rules! init_progress {
-    ($local:expr, $label:expr) => {{
-        let pb = indicatif::ProgressBar::new($local.len() as u64);
-        let mut template =
-            "{spinner:.green} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {pos}/{len} ".to_string();
-        template += $label;
-        template += " ({eta})";
-        pb.set_style(
-            indicatif::ProgressStyle::with_template(&template)
-                .unwrap()
-                .with_key("eta", $crate::utils::eta_key)
-                .progress_chars("#>-"),
-        );
-        pb
-    }};
-}
-
-#[macro_export]
-macro_rules! update_progress {
-    ($pb:ident, $index:expr) => {
-        $pb.set_position(($index + 1) as u64);
-    };
+pub fn init_progress(len: u64, label: &str) -> indicatif::ProgressBar {
+    let pb = indicatif::ProgressBar::new(len);
+    let mut template =
+        "{prefix}{spinner:.green} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {pos}/{len} "
+            .to_string();
+    write!(template, "{}", label).unwrap();
+    template += " ({eta})";
+    pb.set_style(
+        indicatif::ProgressStyle::with_template(&template)
+            .unwrap()
+            .with_key("eta", crate::utils::eta_key)
+            .progress_chars("#>-"),
+    );
+    pb
 }
 
 /// True if the network calculates gas costs differently.
