+++ crates/script/src/receipts.rs
@@ -1,16 +1,12 @@
-use super::sequence::ScriptSequence;
-use crate::progress::{ScriptProgress, SequenceProgress};
 use alloy_chains::Chain;
 use alloy_primitives::{utils::format_units, TxHash, U256};
 use alloy_provider::{PendingTransactionBuilder, Provider};
 use alloy_rpc_types::AnyTransactionReceipt;
 use eyre::Result;
 use foundry_common::provider::RetryProvider;
-use futures::StreamExt;
-use std::sync::Arc;
 
 /// Convenience enum for internal signalling of transaction status
-enum TxStatus {
+pub enum TxStatus {
     Dropped,
     Success(AnyTransactionReceipt),
     Revert(AnyTransactionReceipt),
@@ -26,105 +22,9 @@ impl From<AnyTransactionReceipt> for TxStatus {
     }
 }
 
-/// Gets the receipts of previously pending transactions, or removes them from
-/// the deploy sequence's pending vector
-pub async fn wait_for_pending(
-    provider: Arc<RetryProvider>,
-    sequence_idx: usize,
-    deployment_sequence: &mut ScriptSequence,
-    progress: &ScriptProgress,
-) -> Result<()> {
-    if deployment_sequence.pending.is_empty() {
-        return Ok(());
-    }
-
-    let progress = progress.get_sequence_progress(sequence_idx, deployment_sequence);
-    progress.inner.write().set_status("Checking previously pending transactions");
-    clear_pendings(provider, deployment_sequence, None, &progress).await
-}
-
-/// Traverses a set of pendings and either finds receipts, or clears them from
-/// the deployment sequence.
-///
-/// If no `tx_hashes` are provided, then `deployment_sequence.pending` will be
-/// used. For each `tx_hash`, we check if it has confirmed. If it has
-/// confirmed, we push the receipt (if successful) or push an error (if
-/// revert). If the transaction has not confirmed, but can be found in the
-/// node's mempool, we wait for its receipt to be available. If the transaction
-/// has not confirmed, and cannot be found in the mempool, we remove it from
-/// the `deploy_sequence.pending` vector so that it will be rebroadcast in
-/// later steps.
-pub async fn clear_pendings(
-    provider: Arc<RetryProvider>,
-    deployment_sequence: &mut ScriptSequence,
-    tx_hashes: Option<Vec<TxHash>>,
-    progress: &SequenceProgress,
-) -> Result<()> {
-    let to_query = tx_hashes.unwrap_or_else(|| deployment_sequence.pending.clone());
-
-    let count = deployment_sequence.pending.len();
-
-    trace!("Checking status of {count} pending transactions");
-
-    let futs = to_query.iter().copied().map(|tx| check_tx_status(&provider, tx));
-    let mut tasks = futures::stream::iter(futs).buffer_unordered(10);
-
-    let mut errors: Vec<String> = vec![];
-
-    while let Some((tx_hash, result)) = tasks.next().await {
-        match result {
-            Err(err) => {
-                errors.push(format!("Failure on receiving a receipt for {tx_hash:?}:\n{err}"));
-
-                progress.inner.write().finish_tx_spinner(tx_hash);
-            }
-            Ok(TxStatus::Dropped) => {
-                // We want to remove it from pending so it will be re-broadcast.
-                deployment_sequence.remove_pending(tx_hash);
-                errors.push(format!("Transaction dropped from the mempool: {tx_hash:?}"));
-
-                progress.inner.write().finish_tx_spinner(tx_hash);
-            }
-            Ok(TxStatus::Success(receipt)) => {
-                trace!(tx_hash=?tx_hash, "received tx receipt");
-
-                let msg = format_receipt(deployment_sequence.chain.into(), &receipt);
-                progress.inner.write().finish_tx_spinner_with_msg(tx_hash, &msg)?;
-
-                deployment_sequence.remove_pending(receipt.transaction_hash);
-                deployment_sequence.add_receipt(receipt);
-            }
-            Ok(TxStatus::Revert(receipt)) => {
-                // consider:
-                // if this is not removed from pending, then the script becomes
-                // un-resumable. Is this desirable on reverts?
-                warn!(tx_hash=?tx_hash, "Transaction Failure");
-                deployment_sequence.remove_pending(receipt.transaction_hash);
-
-                let msg = format_receipt(deployment_sequence.chain.into(), &receipt);
-                progress.inner.write().finish_tx_spinner_with_msg(tx_hash, &msg)?;
-
-                errors.push(format!("Transaction Failure: {:?}", receipt.transaction_hash));
-            }
-        }
-    }
-
-    // print any errors
-    if !errors.is_empty() {
-        let mut error_msg = errors.join("\n");
-        if !deployment_sequence.pending.is_empty() {
-            error_msg += "\n\n Add `--resume` to your command to try and continue broadcasting
-    the transactions."
-        }
-        eyre::bail!(error_msg);
-    }
-
-    Ok(())
-}
-
 /// Checks the status of a txhash by first polling for a receipt, then for
 /// mempool inclusion. Returns the tx hash, and a status
-async fn check_tx_status(
+pub async fn check_tx_status(
     provider: &RetryProvider,
     hash: TxHash,
 ) -> (TxHash, Result<TxStatus, eyre::Report>) {
